Introduce mutations control to the 3-way diff strategy
