+++ reconcile/test/fixtures/openshift_resource/deployment.yml
@@ -22,3 +22,10 @@ spec:
         image: nginx:1.14.2
         ports:
         - containerPort: 80
+        resources:
+          requests:
+            memory: 1000Mi
+            cpu: "1"
+          limits:
+            memory: 2000Mi
+            cpu: "2"
